requirements.txt
appdirs==1.4.4
asgiref==3.2.7
beautifulsoup4==4.9.3
boto3==1.17.99
botocore==1.20.99
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
Django==3.0.4
django-allauth==0.39.1
django-crispy-forms==1.7.2
django-filter==2.4.0
django-import-export==2.4.0
django-storages==1.11.1
docopt==0.6.2
et-xmlfile==1.0.1
filelock==3.0.12
googletrans==3.1.0a0
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==3.3.0
jdcal==1.4.1
jmespath==0.10.0
joblib==0.15.1
Js2Py==0.70
MarkupPy==1.14
nltk==3.5
num2words==0.5.10
oauthlib==3.0.1
odfpy==1.4.1
openpyxl==3.0.5
packaging==20.8
Pillow==8.2.0
pipenv==2020.11.15
pipwin==0.5.0
pycparser==2.20
pyjsparser==2.7.1
pyparsing==2.4.7
PyPrind==2.11.2
pySmartDL==1.3.4
python-dateutil==2.8.1
python3-openid==3.1.0
pythonnet==2.5.2
pytz==2019.3
pywebview==3.4
PyYAML==5.3.1
regex==2020.6.8
requests==2.25.1
requests-oauthlib==1.2.0
rfc3986==1.4.0
s3transfer==0.4.2
six==1.15.0
sniffio==1.2.0
soupsieve==2.1
SpeechRecognition==3.8.1
sqlparse==0.3.1
tablib==3.0.0
tqdm==4.46.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4
virtualenv==20.2.2
virtualenv-clone==0.5.4
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.0
